+++ CHANGELOG.md
@@ -5,7 +5,8 @@ S-Parameter Viewer Changelog
 0.28b2 (2025-03-18)
 -------------------
 
-- new: method `map()`, which allows to apply arbitrary NumPy functions to S-parameters
+- new: method `SParams.map()`, which allows to apply arbitrary NumPy functions to S-parameters
+- new: method `SParams.rename()` to rename a trace
 
 
 0.28b1 (2025-03-11)